Added budget forecast
Added a "forecast source" selection In lieu of finding a solution for the schedules vs. budget double counting problem.

Added stackable bars, for past/future separation in the same month.

Some WIP code in there (goalQuery) to be possibly removed once I remember what I was trying to achieve.

and linting to be completed...

Filtering is still permitted in "budget mode" but some filtering scenarios may not make much sense.